+++ adapters/Meta/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 #### Next version
 - Removed the deprecated willBackgroundApplication delegate methods from the banner and interstitial renderers.
+- Replaced the deprecated method GADMobileAds.sharedInstance.sdkVersion with GADMobileAds.sharedInstance.versionNumber.
 
 #### [Version 6.14.0.0](https://dl.google.com/googleadmobadssdk/mediation/ios/meta/MetaAdapter-6.14.0.0.zip)
 - Verified compatibility with Meta Audience Network SDK 6.14.0.